add product list

+++ app/src/main/java/com/yusuforhan/composetest/ProductModel.kt
@@ -3,6 +3,6 @@ package com.yusuforhan.composetest
 data class ProductModel(
     val image : Int,
     val title : String,
-    val rate : Int,
+    val rate : Double,
     val reviewSize : Int
 )
